execute the run block under the same principles than #requires
This way, we allow the block to require and raise exceptions in
files that are "unknown" to autorespawn.

However, this should be discouraged, as it will always cause
autorespawn to execute the block twice. The first time to 'discover'
these files and the second one to add them into its state. This
is needed since at the point of first execution we don't if the
file that is loaded in the interpreter is the same file that is
on disk.

Separating the requires removes this problem