unset the initial state envvar on initialization, and avoid using STDIN as our channel
This fixes a bug if a process uses autorespawn and spawns a new
process that itself is using autorespawn, when the two autorespawn
instances are not related. Before this commit, the child would
get the inherited envvar and assume that it is respawned from another
instance (which is not the case).