expand the set of hooks available to track execution progress
This adds the hooks gem to implement the hooks. It adds a bunch
of hooks to Manager to track the new/removed slaves as well as their
execution. In addition, a 'self slave' is now added to represent the
local process, which makes the association of slaves with PIDs consistent.