use nonblocking writes to push the initial state to the slaves in Manager
Pipes have notoriously small buffers, which meant that #spawn has to
wait until the slave is up and operational before it returns. This
commit allows the caller to #spawn to handle the sending by calling
write_initial_dump repeatedly, and Manager uses that facility by default.

The default behaviour of #spawn is unchanged.